Umbenennen von RestControllerImpl & Auslagern der Funktionen
